Added author annotations to the docs.
