+++ src/main/scala/org/economicsl/auctions/multiunit/LimitAskOrder.scala
@@ -20,7 +20,16 @@ import java.util.UUID
 import org.economicsl.auctions.{Price, Quantity, SinglePricePoint, Tradable}
 
 
-/** An order to sell multiple units of a tradable at a per-unit price greater than or equal to the limit price. */
+/** An order to sell multiple units of a tradable at a per-unit price greater than or equal to the limit price.
+  *
+  * @param issuer
+  * @param limit
+  * @param quantity
+  * @param tradable
+  * @tparam T the type of `Tradable` for which the `Order` is being issued.
+  * @author davidrpugh
+  * @since 0.1.0           
+  */
 class LimitAskOrder[+T <: Tradable](val issuer: UUID, val limit: Price, val quantity: Quantity, val tradable: T)
   extends AskOrder[T] with SinglePricePoint[T]
 
@@ -28,6 +37,9 @@ class LimitAskOrder[+T <: Tradable](val issuer: UUID, val limit: Price, val quan
 /** Companion object for `LimitAskOrder`.
   *
   * Provides default ordering for the multi-unit `LimitAskOrder` type.
+  *
+  * @author davidrpugh
+  * @since 0.1.0
   */
 object LimitAskOrder {
