+++ src/main/scala/org/economicsl/auctions/multiunit/MarketAskOrder.scala
@@ -20,7 +20,15 @@ import java.util.UUID
 import org.economicsl.auctions.{Price, Quantity, SinglePricePoint, Tradable}
 
 
-/** An order to sell a multiple units of a tradable at any positive price. */
+/** An order to sell a multiple units of a tradable at any positive price.
+  *
+  * @param issuer
+  * @param quantity
+  * @param tradable
+  * @tparam T the type of `Tradable` for which the `Order` is being issued.
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 class MarketAskOrder[+T <: Tradable](val issuer: UUID, val quantity: Quantity, val tradable: T)
   extends AskOrder[T] with SinglePricePoint[T] {
 
@@ -29,6 +37,11 @@ class MarketAskOrder[+T <: Tradable](val issuer: UUID, val quantity: Quantity, v
 }
 
 
+/** Companion object for `MarketAskOrder`.
+  *
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 object MarketAskOrder {
 
   def apply[T <: Tradable](issuer: UUID, quantity: Quantity, tradable: T): MarketAskOrder[T] = {
