--- src/main/scala/org/economicsl/auctions/quotes/PriceQuoteRequest.scala
@@ -1,32 +1,0 @@
-/*
-Copyright (c) 2017 KAPSARC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package org.economicsl.auctions.quotes
-
-
-/** Base trait for all price quote requests. */
-sealed trait PriceQuoteRequest extends QuoteRequest
-
-
-/** Used by auction participants to request the current ask price quote. */
-class AskPriceQuoteRequest extends PriceQuoteRequest
-
-
-/** Used by auction participants to request the current bid price quote. */
-class BidPriceQuoteRequest extends PriceQuoteRequest
-
-
-/** Used by auction participants to request the current spread quote. */
-class SpreadQuoteRequest extends PriceQuoteRequest
