+++ src/main/scala/org/economicsl/auctions/quotes/Quote.scala
@@ -18,16 +18,50 @@ package org.economicsl.auctions.quotes
 import org.economicsl.auctions.Price
 
 
+/** Base trait for all quote implementations.
+  *
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 sealed trait Quote
 
+
+/** Base trait for all price quote implementations.
+  *
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 trait PriceQuote extends Quote {
 
   def quote: Price
 
 }
 
+
+/** Class implementing an ask price quote.
+  *
+  * @param quote
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 case class AskPriceQuote(quote: Price) extends PriceQuote
 
+
+/** Class implementing a bid price quote.
+  *
+  * @param quote
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 case class BidPriceQuote(quote: Price) extends PriceQuote
 
+
+/** Class implementing a spread quote.
+  *
+  * @param quote
+  * @author davidrpugh
+  * @since 0.1.0
+  * @todo a spread is the difference between two prices and as such the unit is not really `Price` but rather should
+  *       be `Currency`.
+  */
 case class SpreadQuote(quote: Price) extends PriceQuote
