+++ src/main/scala/org/economicsl/auctions/singleunit/Fill.scala
@@ -20,7 +20,15 @@ import java.util.UUID
 import org.economicsl.auctions.{Contract, Price, Tradable}
 
 
-/** Note that a Fill is also a type of Contract! */
+/** Class representing a `Fill`.
+  *
+  * @param askOrder
+  * @param bidOrder
+  * @param price
+  * @tparam T the type of `Tradable` for which the `Fill` is being issued.
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 case class Fill[T <: Tradable](askOrder: AskOrder[T], bidOrder: BidOrder[T], price: Price) extends Contract {
 
   /** By convention a `Fill` represents a liability of the buyer */
