+++ src/main/scala/org/economicsl/auctions/singleunit/MarketBidOrder.scala
@@ -20,7 +20,14 @@ import java.util.UUID
 import org.economicsl.auctions.{Price, Tradable}
 
 
-/** An order to buy a single-unit of a tradable at any positive price. */
+/** An order to buy a single-unit of a `Tradable` at any positive price.
+  *
+  * @param issuer
+  * @param tradable
+  * @tparam T the type of `Tradable` for which the `MarketBidOrder` is being issued.
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 class MarketBidOrder[+T <: Tradable](val issuer: UUID, val tradable: T) extends BidOrder[T] {
 
   val limit: Price = Price.MaxValue
@@ -28,6 +35,11 @@ class MarketBidOrder[+T <: Tradable](val issuer: UUID, val tradable: T) extends 
 }
 
 
+/** Companion object for `MarketBidOrder`.
+  *
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 object MarketBidOrder {
 
   def apply[T <: Tradable](issuer: UUID, tradable: T): MarketBidOrder[T] = {
