+++ src/main/scala/org/economicsl/auctions/singleunit/Order.scala
@@ -18,12 +18,21 @@ package org.economicsl.auctions.singleunit
 import org.economicsl.auctions.{Contract, OrderLike, SingleUnit, Tradable}
 
 
+/** Base trait for all single-unit orders.
+  *
+  * @tparam T the type of `Tradable` for which the `Order` is being issued.
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 sealed trait Order[+T <: Tradable] extends Contract with OrderLike[T] with SingleUnit[T]
 
 
 /** Companion object for the `Order` trait.
   *
   * Defines a basic ordering for anything that mixes in the `Order` trait.
+  *
+  * @author davidrpugh
+  * @since 0.1.0
   */
 object Order {
 
@@ -39,7 +48,19 @@ object Order {
 }
 
 
+/** Base trait for all single-unit orders to sell a particular `Tradable`.
+  *
+  * @tparam T the type of `Tradable` for which the `AskOrder` is being issued.
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 trait AskOrder[+T <: Tradable] extends Order[T]
 
 
+/** Base trait for all single-unit orders to buy a particular `Tradable`.
+  *
+  * @tparam T the type of `Tradable` for which the `BidOrder` is being issued.
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 trait BidOrder[+T <: Tradable] extends Order[T]
