+++ src/main/scala/org/economicsl/auctions/singleunit/pricing/BidQuotePricingPolicy.scala
@@ -23,6 +23,8 @@ import org.economicsl.auctions.{Price, Tradable}
   *
   * @note (M+1)th highest price is equivalent to the bid quote. It is incentive compatible for buyers to truthfully
   *       reveal their respective valuations in single-unit auctions using this pricing rule.
+  * @author davidrpugh
+  * @since 0.1.0
   */
 class BidQuotePricingPolicy[T <: Tradable] extends PricingPolicy[T] {
