+++ src/main/scala/org/economicsl/auctions/singleunit/twosided/ClearResult.scala
@@ -19,5 +19,14 @@ import org.economicsl.auctions.Tradable
 import org.economicsl.auctions.singleunit.Fill
 
 
+/** Class representing the results of clearing a `DoubleAuction`.
+  *
+  * @param fills
+  * @param residual
+  * @tparam T
+  * @tparam A
+  * @author davidrpugh
+  * @since 0.1.0
+  */
 case class ClearResult[T <: Tradable, +A <: DoubleAuction[T]](fills: Option[Stream[Fill[T]]], residual: A)
